Load google-code-prettify from web
Load google-code-prettify (syntax highlighting for assignment write up)
from the web, not from local javascript files. See
https://github.com/google/code-prettify/blob/master/docs/getting_started.md

Motivation: Fewer files for students to deal with. Also, this is simpler
and always uses the most up-to-date version of google-code prettify.
Drawbacks: You need a network connection to get syntax highlighting in
the assignment writeup.